requirements.txt
attrs==25.3.0
brother-ql==0.9.4
click==8.2.1
future==1.0.0
packbits==0.6
Pillow==9.5.0
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.1.0
pyusb==1.3.1
rsa==4.9.1
Telethon==1.40.0
